src/main/java/bg/softuni/regular_exam/controllers/GlobalExceptionHandler.java
package bg.softuni.regular_exam.controllers;

import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.servlet.ModelAndView;

@ControllerAdvice
public class GlobalExceptionHandler {
    @ExceptionHandler( ForbiddenException.class)
    public static ModelAndView handleDatabaseErrors(ForbiddenException e) {
        ModelAndView modelAndView = new ModelAndView("error");
        modelAndView.addObject("message", e.getMessage());

        return modelAndView;
    }}


